refactor: INV4 changes
