+++ runtime/tinkernet/Cargo.toml
@@ -26,12 +26,15 @@ getrandom = { version = "0.2.4", default-features = false, features = ["js"] }
 
 # InvArch Pallets
 #pallet-ipf = { git = "https://github.com/InvArch/InvArch-Frames", rev = "590777fa17e8e94c5620265a12c4aff5104b2d5c", default-features = false }
-pallet-inv4 = { git = "https://github.com/InvArch/InvArch-Frames", branch = "gabriel-rmrk", default-features = false }
+#pallet-inv4 = { git = "https://github.com/InvArch/InvArch-Frames", branch = "gabriel-rmrk", default-features = false }
 #pallet-ipf = { path = "../../../InvArch-Pallet-Library/INV4/pallet-ipf", default-features = false }
 
-#pallet-inv4 = { path = "../../../InvArch-Frames/INV4/pallet-inv4", default-features = false }
+pallet-inv4 = { path = "../../../InvArch-Frames/INV4/pallet-inv4", default-features = false }
 
 pallet-maintenance-mode = { path = "../../pallets/maintenance-mode", default-features = false }
+
+pallet-referenda = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.36" }
+once_cell = { version = "1.17.0", default-features = false }
 
 ## Parity's Unique + RMRK Pallets
 pallet-rmrk-core = { git = "https://github.com/rmrk-team/rmrk-substrate", default-features = false, rev = "9c538c23aa53293691833b80e890e67aa524b062" }
